fix(core): satisfy changed-file analyzers
