+++ src/ModularPipelines/Engine/ModuleExecutionPipeline.cs
@@ -15,22 +15,6 @@ using Polly;
 namespace ModularPipelines.Engine;
 
 /// <summary>
-/// Interface for the module execution pipeline.
-/// </summary>
-internal interface IModuleExecutionPipeline
-{
-    /// <summary>
-    /// Executes a module with all applicable behaviors.
-    /// </summary>
-    /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-    Task<ModuleResult<T>> ExecuteAsync<T>(
-        Module<T> module,
-        ModuleExecutionContext<T> executionContext,
-        IModuleContext moduleContext,
-        CancellationToken engineCancellationToken);
-}
-
-/// <summary>
 /// Orchestrates module execution by applying behaviors based on module configuration.
 /// </summary>
 /// <remarks>
@@ -354,11 +338,11 @@ internal class ModuleExecutionPipeline : IModuleExecutionPipeline
                     timeoutException.ElapsedTime.ToDisplayString());
             }
         }
-
         else if (IsTimeout(config, executionContext, exception))
         {
             executionContext.Status = Status.TimedOut;
         }
+
         // Check for pipeline cancellation
         else if (IsPipelineCancelled(exception))
         {
