+++ src/ModularPipelines/Extensions/PipelineBuilderExtensions.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 using ModularPipelines.Engine;
@@ -83,6 +84,7 @@ public static class PipelineBuilderExtensions
     /// <typeparam name="T">Any type from the assembly to scan.</typeparam>
     /// <param name="builder">The pipeline builder.</param>
     /// <returns>The same builder instance for chaining.</returns>
+    [RequiresUnreferencedCode("Module discovery scans all types in an assembly.")]
     public static PipelineBuilder AddModulesFromAssemblyContainingType<T>(this PipelineBuilder builder)
     {
         return builder.AddModulesFromAssembly(typeof(T).Assembly);
@@ -94,6 +96,7 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <param name="assembly">The assembly to scan.</param>
     /// <returns>The same builder instance for chaining.</returns>
+    [RequiresUnreferencedCode("Module discovery scans all types in an assembly.")]
     public static PipelineBuilder AddModulesFromAssembly(this PipelineBuilder builder, Assembly assembly)
     {
         ArgumentNullException.ThrowIfNull(assembly);
@@ -134,7 +137,9 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TRequirement">The type of requirement to add.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddRequirement<TRequirement>(this PipelineBuilder builder)
+    public static PipelineBuilder AddRequirement<
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TRequirement>(
+        this PipelineBuilder builder)
         where TRequirement : class, IPipelineRequirement
     {
         builder.Services.AddRequirement<TRequirement>();
@@ -163,7 +168,9 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TGlobalSetup">The type of hook class.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddPipelineGlobalHooks<TGlobalSetup>(this PipelineBuilder builder)
+    public static PipelineBuilder AddPipelineGlobalHooks<
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TGlobalSetup>(
+        this PipelineBuilder builder)
         where TGlobalSetup : class, IPipelineGlobalHooks
     {
         builder.Services.AddPipelineGlobalHooks<TGlobalSetup>();
@@ -176,7 +183,9 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TReceiver">The receiver type.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddModuleEventReceiver<TReceiver>(this PipelineBuilder builder)
+    public static PipelineBuilder AddModuleEventReceiver<
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TReceiver>(
+        this PipelineBuilder builder)
         where TReceiver : class, IModuleEventReceiver
     {
         builder.Services.AddModuleEventReceiver<TReceiver>();
@@ -226,7 +235,7 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TRepository">The type of result repository to add.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddResultsRepository<TRepository>(this PipelineBuilder builder)
+    public static PipelineBuilder AddResultsRepository<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TRepository>(this PipelineBuilder builder)
         where TRepository : class, IModuleResultRepository
     {
         builder.Services.AddSingleton<IModuleResultRepository, TRepository>();
@@ -274,7 +283,7 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TProvider">The type of estimated time provider to add.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddModuleEstimatedTimeProvider<TProvider>(this PipelineBuilder builder)
+    public static PipelineBuilder AddModuleEstimatedTimeProvider<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TProvider>(this PipelineBuilder builder)
         where TProvider : class, IModuleEstimatedTimeProvider
     {
         builder.Services.AddSingleton<IModuleEstimatedTimeProvider, TProvider>();
@@ -287,7 +296,9 @@ public static class PipelineBuilderExtensions
     /// <param name="builder">The pipeline builder.</param>
     /// <typeparam name="TWriter">The type of pipeline file writer to add.</typeparam>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddPipelineFileWriter<TWriter>(this PipelineBuilder builder)
+    public static PipelineBuilder AddPipelineFileWriter<
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TWriter>(
+        this PipelineBuilder builder)
         where TWriter : class, IBuildSystemPipelineFileWriter
     {
         builder.Services.AddSingleton<IBuildSystemPipelineFileWriter, TWriter>();
@@ -313,7 +324,10 @@ public static class PipelineBuilderExtensions
     /// <typeparam name="TImplementation">The implementation type.</typeparam>
     /// <param name="builder">The pipeline builder.</param>
     /// <returns>The same builder instance for chaining.</returns>
-    public static PipelineBuilder AddSingleton<TService, TImplementation>(this PipelineBuilder builder)
+    public static PipelineBuilder AddSingleton<
+        TService,
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TImplementation>(
+        this PipelineBuilder builder)
         where TService : class
         where TImplementation : class, TService
     {
